requirements.txt
FLASK==1.1.2
requests==2.24.0
Pillow==7.2.0
azure-cosmos==4.0.0